chore(ds-app): organise SideNavigation story helpers under #storybook/navigation
Group the story-only helpers into a navigation domain folder and expose them via
a package subpath import, replacing deep-relative paths.

- Move story-utils + fixtures to src/storybook/navigation/{story-utils,fixtures}.tsx
  (the storybook folder is excluded from the package build, so none of this ships).
- Add a `#storybook/*` subpath to package.json `imports`; stories now import from
  `#storybook/navigation/...` instead of `../../../../storybook/...`. Resolves
  with no extra config (NodeNext + Vite/Storybook read package.json imports).
- Fixtures use the shared `MockBadge` (single badge definition) for leaf `slot`s.
- Document the `#`-alias convention in root AGENTS.md so it isn't re-derived.
