+++ packages/react/ds-app/package.json
@@ -4,6 +4,9 @@
   "type": "module",
   "module": "dist/esm/index.js",
   "types": "dist/types/index.d.ts",
+  "imports": {
+    "#storybook/*": "./src/storybook/*"
+  },
   "files": [
     "dist"
   ],
